Import and modify the icontrol gem into the module.
The original icontrol gem relies on the built in ruby SOAP stuff to do
it's magic but this was removed in ruby 1.9.  It's possible to add back
in but it's easier to switch to savon which works across 1.8-2.1.

The icontrol gem is basically an extremely thin wrapper around the wsdl
files with a couple of methods to initialize the SOAP connections, so
we're able to easily update it to use Savon.